+++ github-page/changelog.md
@@ -35,6 +35,7 @@ redirect_from:
 - Change warning message for `warnMixed`
 - If an SchemaType dosnt extend `SchemaTypeOptions`, the options are now defaulted to the outer-layer
 - `innerOptions` and `outerOptions` can now be used for Maps too
+- Custom Validators now support `message` being an function
 - [IC] `mapOptions` now always errors if the given type dosnt extend `mongoose.SchemaTypeOptions`
 - [IC] Moved function `_buildPropMetadata` to its own file (`processProp`) and renamed it to `processProp` [{% include gitissue repo="typegoose" num=286 %}]
 - [IC] Moved pre-processing in function `prop` into `processProp` [{% include gitissue repo="typegoose" num=286 %}]